pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.21.0",
    "discord-py>=2.5.2",
    "flask>=3.1.1",
    "pynacl>=1.5.0",
    "yt-dlp>=2025.7.21",
]
